+++ auv_control_msgs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for package auv_control_msgs
 
+## 0.2.0 (2025-04-28)
+
+- Implements the EndEffectorTrajectoryPoint and EndEffectorTrajectory messages
+to support the new end effector trajectory controller
+
 ## 0.1.0 (2025-04-27)
 
 - Implements the IKControllerStateStamped message to support the new IK